backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
mysql-connector-python==8.2.0
chromadb==0.4.18
sentence-transformers==2.2.2
python-docx==1.1.0
PyPDF2==3.0.1
pymupdf==1.23.8
pdfplumber==0.10.3
requests==2.31.0
python-dotenv==1.0.0
tenacity==8.2.3